requirements.txt
asgiref==3.7.2
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Django==5.0.2
djangorestframework==3.14.0
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
greenlet==3.0.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.10.3
platformdirs==2.6.2
Pygments==2.17.2
python-dotenv==0.21.1
pytz==2024.1
requests==2.28.2
SQLAlchemy==2.0.4
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2024.1
urllib3==1.26.14
Werkzeug==2.2.3
